moving the extraction of original exception properties to the util module
